Move fmp4 utilities to fmp4utils.py
